fix: Bug on stoplight list of strings (#1984)
There is a really odd css bug with `@stoplight/elements` which is used to render the openapi specs.
When it renders a list of strings, every string renders in columns (see screenshot below, hard to describe)

The problem is that the container of the list's classname get prepend with an additional `-wrap` which is `sl-flex-wrap-wrap` instead of just `sl-flex-wrap`.

Cannot say why it's happening, so as a quick fix I add the missing css for that classname.