feat(seo): Add better metadata to pages (#2399)
Adding a `SEO` component that adds a page `description`, `title`, and `pathname` in the `<head>` section. This should allow better ranking in search engine for the docs.

I've put a default description in the gatsby's config because it needed one. But a page description is the one in each mdx file metadata