fix: always add trailing Slash (#2444)
Using `trailingSlash: 'ignore'` will not generates redirect with page ending or not by `/`

For Cloudflare, if no page found, `/foo` == `/foo/` == `/foo/index.html`. And no redirect occurs foo/index.html will be serve for all of them.

We need to be explicit and choose which type of URL we prefer, and Gatsby will generate the appropriate pages/code to redirect all others to our prefered one.

So we need to set `trailingSlash` to `always` or `never`.
I pick `always` like we did for the old documentation.

Fixes MRGFY-2702