feat(seo): Generate og:image for each pages (#2445)
We now generate images for each pages, this image serve as thumbnail when shared on social media.

Each image is generated by drawing the page's title on a pre-made template. The drawing is achieve using `canvas` module which is basically the HTML's `canvas` API running in node.js.

Each image is generated at build time , and is stored in the `static` folder, which makes them available for robots to be crawled based on `og:image` meta tag on each page.