fix(algolia): Replace `gatsby-plugin-algolia` by own integration (#2530)
Replacing `gatsby-plugin-algolia` for indexing on algolia by our own plugin. The previous one had issues updating the index: when you update the structure of a record it won't update it because it is based on an id which change only if the page's content changes.

Now we use `saveObjects` from algolia API which does a delete and replace, which makes more sense for us.