fix: Anchor link is hard to click on (#3296)
Left sidebar was on top of the anchor link, without visually hiding it. But the hover was stoped by the sidebar midway on the icon making it disapear and hard to click on.
Also increasing the anchor size to match the title font size, making it a little bit easier to click on.