fix(api): replace vue component by prebuild js (#3620)
We can't find what wrong with the @scalar vue component we use.
The issue is somewhere into a sub dependency of @scalar/api-reference.
NPM does not provide an easy way to cap version of subdependencies.

This change replace the component with the already built JS file to
load our API docs.

Fixes INC-141