refactor(schema): Remove $id from all definitions (#3755)
Putting `$id` on every definitions can be hard to handle when we try to reference those. Because `$id` serve as base path when resolving from a schema.

They were used only to get links to some specific types pages, and can be retrieve by using the last segment of the URI.